[skip ci] Add CI testing pipelines
CI testing implies link to available manticore code.
This code will be populated, built and then test run over it.
Without available manticore testing is impossible.

Manticore have to be achieved via locator, which can be defined one of the following methods (from first to last):

  1. as environment 'MANTICORE_LOCATOR'
  2. as content of the file 'local_manticore_src.txt'
  3. as content of the file 'manticore_src.txt'

 Format of the locator is one passed to cmake's ExternalProject. See https://cmake.org/cmake/help/latest/module/ExternalProject.html#command:externalproject_add for details.

 Environment read first, if it exists, it will be used. That is suitable for IDE and kind of cross-testing, where same sources shared, but placement of manticore may be different among hosts.

 Then files `local_manticore_src.txt` and `manticore_src.txt` First one is added to .gitignore, so is local. Last one is commited and bring 'production link'.

This very commit marked by 'skip ci' because it doesn't yet linked to a manticore revision.
Linkage is critical, since columnar may be built without manticore (it is necessary only for testing), but manticore can't be built without columnar (as it uses headers to know how to compile).

So, this commit being pushed, provides it's hash for manticore to link with.
Then we can also make back-link here to that version, for testing. So, no circular dependency here.