ci: run CLT tests related to MCL automatically
The new feature runs CLT tests from the daemon repository using the current MCL code automatically in test.yml. If a branch with the same name exists in the daemon repo, the tests and the base test kit are taken from that branch. Otherwise, the master branch and the default test kit are used.

Related PR: https://github.com/manticoresoftware/columnar/pull/122